Rewrite soynode to use OO instead of global state
